Make shared_ptr and luabind work together for VS11.
Because luabind tries to find get_pointer by ADL, it does matter if shared_ptr is
actually defined in std::tr1 and only brought to std by the using keyword.
VS11 does the right thing and defines it directly in std.

This implementation of get_pointer is #ifdefed, to work on both VS10 and 11.