Work around crash in lua_close.
The order in which finalizers are called is undefined for lua_close. So efforts are made
to collect as much as possible before (clear _G, package.loaded and package.preload and
do a full GC before and after).
In addition, the most (only?) problematic function wrap_Component::cleanupComponent() now
checkes if the lua_State is closing and issues a warning in this case, instead of
dispatching to Lua (and chrashing).