container lexp: optionally return references (get).
By setting the boolean 'ref' template argument to export*Container() to true,
*Container_get() returns references to instead of copies of the containers
elements.
This leaves the problem of Lua being able to hold an invalid reference to
Elements of a destroyed container. Normally, the dependency_policy should be
used here, but it turned out that Luabind does not support more than
'sizeof(object_rep)' dependencies per object. If dependencies are created in
a loop and the GC does not run, the limit is quicly exceeded.