Add partial unicode support.
This applies to Windows mostly. Any system which uses UTF-8 as its narrow
encoding should be fine.

Previously the game even failed to start if the program file was in a directory
containing non-ASCII characters. This is fixed now, jd is now fully unicode
transparent at this end.

However, since the commandline arguments are received in local 8-bit encoding
and the logfile is opened via the C++ iostreams API which also works with this
encoding, only characters which are representable in this encoding are
supported in game (file-)names.

The UCS2/uint16_t vs. wchar_t distinction is neccessary because on many systems
wchar_t has only one byte, just like an ordinary char. On these, wstrings are
UTF-8 encoded.